Change lockWatcher in trackedlock to list 8 locks held longest.
Change the lockWatcher in the trackedlock to report the 8
locks that were held the longest and not just 1 (controlled by
globals.lockWatcherLocksLogged which is not tunable, but could
be changed via gdb).

Seeing more locks should be useful in debugging deadlocks (its likely
that the lock the held second longest is the lock that the thread holding
the first lock wants) and may also help in those cases where the lock
held longest is not the cause of the deadlock (perhaps because a thread
got the lock and did some time-consuming work before it then tried to
acquire the lock held by the thread causing the deadlock.

Also try to minimize the chance of hitting a SIGSEGV that can occur if
we try to copy the lock stack of a Mutex that has been held too long
and it happens to be unlocked right then.

The change to log multiple locks held too long required changes in the
tests to look at the new "rank %d" log field.

Tighten up the log entry parser logger.ParseLogEntry() a bit.  Have it
log a warning if a log entyr doesn't match its RE.  Fix some bugs in
swiftclient tests that were exposed by this.

Improve comments in this and related code.  Improve the trackedlock
log messages as well.