Performed the simple steps in inode/api_snapshot_changes_needed.txt
Also switched dirty state to FALSE for paged-in Inodes. It's still
unclear why the inMemoryInode, when configured after fetching the
onDiskInodeV1 from headhunter, was marked DIRTY... it should have
been CLEAN. Any operation that subsequently modifies the inode must
already be setting the flag to DIRTY as they do it, so there was
literally no benefit to setting the flag to DIRTY upon load.